backups/post_conversation_fix_20250628_200437/requirements.txt
# ========================================
# TELEGRAM BOT - ПОЛНЫЕ ЗАВИСИМОСТИ
# ========================================
# Дата обновления: 8 июня 2025
# Python версия: 3.12.3
# Операционная система: Ubuntu 24.04.2 LTS

# ========================================
# ОСНОВНЫЕ ЗАВИСИМОСТИ БОТА
# ========================================

# Telegram Bot API - основная библиотека для работы с Telegram
python-telegram-bot>=22.1

# HTTP клиенты для API запросов  
httpx>=0.28.1
aiohttp>=3.12.10

# Конфигурация и переменные окружения
python-dotenv>=1.1.0

# Работа с YAML файлами (конфигурация полей)
pyyaml>=6.0.2

# ========================================
# ИНТЕГРАЦИИ С ВНЕШНИМИ СЕРВИСАМИ
# ========================================

# Google Sheets интеграция
gspread>=6.2.1
google-auth>=2.40.3
google-auth-oauthlib>=1.2.2
oauth2client>=4.1.3

# ========================================
# СИСТЕМНЫЕ ЗАВИСИМОСТИ
# ========================================

# HTTP библиотеки
requests>=2.32.3
httplib2>=0.22.0

# Асинхронность
aiohappyeyeballs>=2.6.1
aiosignal>=1.3.2
anyio>=4.9.0

# Криптография и безопасность
certifi>=2025.4.26
rsa>=4.9.1

# Парсинг и валидация
attrs>=25.3.0
charset-normalizer>=3.4.2
idna>=3.10
pyparsing>=3.2.3

# Типизация
typing_extensions>=4.14.0

# ========================================
# ДОПОЛНИТЕЛЬНЫЕ УТИЛИТЫ  
# ========================================

# Кэширование
cachetools>=5.5.2

# Работа с URL и многозначными словарями
yarl>=1.20.0
multidict>=6.4.4
frozenlist>=1.6.2
propcache>=0.3.1

# OAuth аутентификация
oauthlib>=3.2.2
requests-oauthlib>=2.0.0

# ASN.1 поддержка для Google API
pyasn1>=0.6.1
pyasn1_modules>=0.4.2

# HTTP/1.1 support
h11>=0.16.0
httpcore>=1.0.9

# Network detection
sniffio>=1.3.1

# Утилиты
six>=1.17.0
urllib3>=2.4.0